+++ .env.example
@@ -14,3 +14,6 @@ CHECK_TIMETABLE_URL=http://www.teu.ac.jp/campus/access/006644.html#bustimetable
 
 # Slack Notifier
 WEBHOOK_URL=
+
+# Self Hostname
+SELF_HOSTNAME=
